Rename files using camelcase for consistency
