+++ routes/userWebRoutes.js
@@ -1,17 +1,15 @@
 const express = require('express');
-const routingController = require('../controllers/api_routing_controller');
-
+const routingController = require('../controllers/webRoutingController');
 const validator = require('../middleware/validator.js');
 const createValidationRules = validator.createValidationRules;
 const retrieveValidationRules = validator.retrieveValidationRules;
+const updateValidationRules = validator.updateValidationRules;
 const deleteValidationRules = validator.deleteValidationRules;
-const changeValidationRules = validator.changeValidationRules;
-const apiValidate = validator.apiValidate;
-
+const webValidate = validator.webValidate;
 const converter = require('../middleware/handleId.js').convertId;
 
-const userAPIRouter = express.Router();
-module.exports = userAPIRouter;
+const userWebRouter = express.Router();
+module.exports = userWebRouter;
 
 /**
  * @api {get} /list Request list of users
@@ -21,7 +19,7 @@ module.exports = userAPIRouter;
  *
  * @apiSuccess {JSON object} JSON object of users.
  */
-userAPIRouter.get('/user/list', routingController.getUsersList);
+userWebRouter.get('/user/list', routingController.getUsersList);
 
 /**
  * @api {get} /user/:userId Request single user
@@ -31,27 +29,27 @@ userAPIRouter.get('/user/list', routingController.getUsersList);
  *
  * @apiSuccess {JSON object} JSON object of users.
  */
-userAPIRouter.get('/user/:userId', retrieveValidationRules(), apiValidate, routingController.retrieveUser);
+userWebRouter.get('/user/:userId', converter, retrieveValidationRules(), webValidate, routingController.retrieveUser);
 
 /**
- * @api {post} /enroll   single user
- * @apiName EnrollUser
+ * @api {post} /enroll/   single user
+ * @apiName GetUserById
  *
  * @apiParam {String} userId User's unique ID
  *
  * @apiSuccess {JSON object} JSON object of users.
- **/
-userAPIRouter.post('/user/enroll', converter, createValidationRules(), apiValidate, routingController.enrollUser);
+ */
+userWebRouter.post('/user/enroll', converter, createValidationRules(), webValidate, routingController.enrollUser);
 
 /**
- * @api {post} /update  single user
- * @apiName UpdateUser
+ * @api {post} /update   single user
+ * @apiName GetUserById
  *
- * @apiParam {String} userId User's unique ID and fields to be updated
+ * @apiParam {String} userId User's unique ID
  *
  * @apiSuccess {JSON object} JSON object of users.
  */
-userAPIRouter.post('/user/update', converter, changeValidationRules(), apiValidate, routingController.updateUser);
+userWebRouter.post('/user/update', converter,  updateValidationRules(), webValidate, routingController.updateUser);
 
 /**
  * @api {post} /remove   single user
@@ -61,4 +59,4 @@ userAPIRouter.post('/user/update', converter, changeValidationRules(), apiValida
  *
  * @apiSuccess {String} message confirming the deletion.
  */
-userAPIRouter.post('/user/remove', converter, deleteValidationRules(), apiValidate, routingController.deleteUser);
+userWebRouter.post('/user/remove', converter, deleteValidationRules(), webValidate, routingController.deleteUser);
